You can now upload and see you pictures in ProfileView, Added FireStorage and ImagePicker ViewModels
