+++ VoyageVaultApp/Views/Authorized/HomeView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 
 struct HomeView: View {
     
+        
     
     @EnvironmentObject var firebaseAuth: FirebaseAuth
